Separate dz types from what it can receive
A hunk should be droppable on a lane, but not vice versa. At the same
time we also allow multiple types on a single drop zone, this lets us
e.g. setup dropzones betweeen files with the same name, while allowing
individual hunks to be droppable onto a new lane.